* ieb/master:
  Fixed shutdown exception in cluster tracking service. The server registration will be removed from other servers by servers left in the cluster, so it doesnt matter if its not there.
  Made the rebundle process aware of BUILD_OPTS as is the clean restart script.
  Fixed some shutdown exceptions with JMS.
  KERN-720 This ensures that UserPostProcessors that are registered after activation of the CreateUserServlet are still invoked for default users. All
  Log exceptions in default-user loop at start-up
  Updated to person directory to match changes in ldap bundle.
  Updated ldap and ldapauth to match reworked results.
  KERN-429 Improvements to the Nightly Build script in hybrid
  KERN-523 Small fixes after review of the spec.
  KERN-523 Sorted out most of workflow persistence in JCR.
  KERN-523 Start of work on a workflow engine and rule persistence, more like research than implementation to uncover the details. (all in contrib)
  KERN-692 Added a sling_install.properties to the bootstrap so that we can add startup properties before OSGi gets up and runnign for the first time. Futher doc on Confluence under a page attached to this jira.
  Removed TODO autogenerated messages
  KERN-742 removed qdoc from version bundle
  KERN-746 remove qdoc from memory bundle

+++ nightly/hybrid/nightly.sh
@@ -4,8 +4,9 @@
 # don't forget to trust the svn certificate permanently: svn info https://source.sakaiproject.org/svn
 # and svn info https://source.caret.cam.ac.uk/camtools
 
-export K2_TAG="0.3"
+export K2_TAG="0.4"
 export S2_TAG="tags/sakai-2.7.0-b06"
+export K2_ARTIFACT="org.sakaiproject.nakamura.app-0.4.jar"
 
 # Treat unset variables as an error when performing parameter expansion
 set -o nounset
@@ -18,15 +19,32 @@ export PATH=$JAVA_HOME/bin:${PATH}
 export MAVEN_HOME=/usr/local/apache-maven-2.2.1
 export M2_HOME=/usr/local/apache-maven-2.2.1
 export PATH=$MAVEN_HOME/bin:${PATH}
-export MAVEN_OPTS="-Xmx1024m -XX:MaxPermSize=512m"
+export MAVEN_OPTS="-Xmx512m -XX:MaxPermSize=256m"
 export JAVA_OPTS="-server -Xmx1024m -XX:MaxPermSize=512m -Djava.awt.headless=true -Dsun.lang.ClassLoader.allowArraySyntax=true -Dsakai.demo=true -Dsakai.cookieName=SAKAI2SESSIONID"
 export K2_OPTS="-server -Xmx512m -XX:MaxPermSize=128m -Djava.awt.headless=true"
 BUILD_DATE=`date "+%D %R"`
 
+# resolve links - $0 may be a softlink
+PRG="$0"
+
+while [ -h "$PRG" ]; do
+    ls=`ls -ld "$PRG"`
+    link=`expr "$ls" : '.*-> \(.*\)$'`
+    if expr "$link" : '/.*' > /dev/null; then
+        PRG="$link"
+    else
+        PRG=`dirname "$PRG"`/"$link"
+    fi
+done
+
+# Get standard environment variables
+PRGDIR=`dirname "$PRG"`
+PRGDIR=`pwd "$PRGDIR"`
+
 # ensure logs directory exists
 if [ ! -d $BUILD_DIR/logs ]
 then
-	mkdir $BUILD_DIR/logs
+	mkdir -p $BUILD_DIR/logs
 fi
 
 # shutdown all running instances
@@ -37,34 +55,34 @@ set -o errexit
 
 # clean previous builds
 cd $BUILD_DIR
-rm -rf sakai
-rm -rf kernel
-rm -rf sakai2-demo
-rm -rf sling
-rm -rf sakai3
-rm -rf ~/.m2/repository/org/sakaiproject
-
-# build sling/trunk
-# not needed as long as hudson builds are getting deployed to apache-snapshot repo
-# org.apache.sling.launchpad.base-2.0.5-SNAPSHOT-app.jar cannot be found - build sling
-#echo "Building sling/trunk..."
-#cd $BUILD_DIR
-#svn export -q http://svn.apache.org/repos/asf/sling/trunk sling
-#cd sling
-#mvn clean install -Dmaven.test.skip=true
-#rm -rf sling
+if [ $1 == "clean" ]
+then
+    echo "Starting clean build..."
+    rm -rf sakai
+    rm -rf sakai2-demo
+    rm -rf sakai3
+    rm -rf ~/.m2/repository/org/sakaiproject
+else
+    echo "Starting incremental build..."
+fi
 
 # build sakai 3
-echo "Building nakamura@$K2_TAG..."
 cd $BUILD_DIR
-mkdir sakai3
+mkdir -p sakai3
 cd sakai3
-git clone -q git://github.com/ieb/open-experiments.git
-cd open-experiments/slingtests/osgikernel/
-git checkout $K2_TAG
-# work around instance not listening on port 80
-#perl -pwi -e 's/localhost/localhost:8080/g' bundles/proxy/src/main/resources/SLING-INF/content/var/proxy/s23/*.json
-mvn clean install -Dmaven.test.skip=true
+if [ -f .lastbuild ]
+then
+    echo "Skipping build nakamura@$K2_TAG..."
+else
+    echo "Building nakamura@$K2_TAG..."
+    git clone -q git://github.com/ieb/open-experiments.git
+    cd open-experiments/slingtests/osgikernel/
+    git checkout -b $K2_TAG
+    # work around instance not listening on port 80
+    #perl -pwi -e 's/localhost/localhost:8080/g' bundles/proxy/src/main/resources/SLING-INF/content/var/proxy/s23/*.json
+    mvn -e clean install -Dmaven.test.skip=true
+    date > .lastbuild
+fi
 
 # start sakai 3 instance
 echo "Starting sakai3 instance..."
@@ -73,58 +91,54 @@ mkdir -p sling/config/org/sakaiproject/nakamura/proxy
 echo 'port=I"8080"' > sling/config/org/sakaiproject/nakamura/proxy/TrustedLoginTokenProxyPreProcessor.config
 echo 'sharedSecret="e2KS54H35j6vS5Z38nK40"' >> sling/config/org/sakaiproject/nakamura/proxy/TrustedLoginTokenProxyPreProcessor.config
 echo 'service.pid="org.sakaiproject.nakamura.proxy.TrustedLoginTokenProxyPreProcessor"' >> sling/config/org/sakaiproject/nakamura/proxy/TrustedLoginTokenProxyPreProcessor.config
-java $K2_OPTS -jar org.sakaiproject.nakamura.app-0.3.jar -p 8008 -f - > $BUILD_DIR/logs/sakai3-run.log.txt 2>&1 &
-
-# untar tomcat
-cd $BUILD_DIR
-tar -xzf apache-tomcat-5.5.26.tar.gz 
-mv apache-tomcat-5.5.26 sakai2-demo
-mkdir sakai2-demo/sakai
+java $K2_OPTS -jar $K2_ARTIFACT -p 8008 -f - > $BUILD_DIR/logs/sakai3-run.log.txt 2>&1 &
 
 # build sakai 2
-echo "Building sakai2/$S2_TAG..."
-svn checkout -q "https://source.sakaiproject.org/svn/sakai/$S2_TAG" sakai
-cd sakai/
-REPO_REV=`svn info|grep Revision`
-# SAK-17223 K2AuthenticationFilter
-rm -rf login/
-svn checkout -q https://source.sakaiproject.org/svn/login/branches/SAK-17223-2.7 login
-# SAK-17222 K2UserDirectoryProvider
-rm -rf providers
-svn checkout -q https://source.sakaiproject.org/svn/providers/branches/SAK-17222-2.7 providers
-# KERN-360 Servlet and TrustedLoginFilter RESTful services
-cp -R $BUILD_DIR/sakai3/open-experiments/hybrid .
-find hybrid -name pom.xml -exec perl -pwi -e 's/2\.8-SNAPSHOT/2\.7-SNAPSHOT/g' {} \;
-perl -pwi -e 's/<\/modules>/<module>hybrid<\/module><\/modules>/gi' pom.xml
-#
-mvn clean install sakai:deploy -Dmaven.test.skip=true -Dmaven.tomcat.home=$BUILD_DIR/sakai2-demo
-cd ..
-
-# configure sakai 2 instance
-cp -f server.xml sakai2-demo/conf/server.xml 
-echo "ui.service = trunk+SAK-17223+KERN-360 on HSQLDB" >> sakai2-demo/sakai/sakai.properties
-echo "version.sakai = $REPO_REV" >> sakai2-demo/sakai/sakai.properties
-echo "version.service = Built: $BUILD_DATE" >> sakai2-demo/sakai/sakai.properties
-#echo "smtp=localhost" >> sakai2-demo/sakai/sakai.properties
-#echo "smtp@org.sakaiproject.email.api.EmailService=localhost" >> sakai2-demo/sakai/sakai.properties
-#echo "smtp.enabled = true" >> sakai2-demo/sakai/sakai.properties
-#echo "smtp.port=8028" >> sakai2-demo/sakai/sakai.properties
-echo "serverName=nightly2.sakaiproject.org" >> sakai2-demo/sakai/sakai.properties
-echo "webservices.allowlogin=true" >> sakai2-demo/sakai/sakai.properties
-echo "webservice.portalsecret=nightly" >> sakai2-demo/sakai/sakai.properties
-echo "samigo.answerUploadRepositoryPath= /tmp/sakai2-hybrid/" >> sakai2-demo/sakai/sakai.properties
-# enable SAK-17223 K2AuthenticationFilter
-echo "top.login=false" >> sakai2-demo/sakai/sakai.properties
-echo "container.login=true" >> sakai2-demo/sakai/sakai.properties
-echo "login.k2.authentication=true" >> sakai2-demo/sakai/sakai.properties
-echo "login.k2.authentication.vaildateUrl=http://localhost:8008/var/cluster/user.cookie.json?c=" >> sakai2-demo/sakai/sakai.properties
-# declare shared secret for trusted login from K2
-echo "org.sakaiproject.util.TrustedLoginFilter.sharedSecret=e2KS54H35j6vS5Z38nK40" >> sakai2-demo/sakai/sakai.properties
+cd $BUILD_DIR
+if [ -f $BUILD_DIR/sakai/.lastbuild ]
+then
+    echo "Skipping build sakai2/$S2_TAG..."
+else
+    echo "Building sakai2/$S2_TAG..."
+    # untar tomcat
+    tar -xzf apache-tomcat-5.5.26.tar.gz 
+    mv apache-tomcat-5.5.26 sakai2-demo
+    mkdir -p sakai2-demo/sakai
+    svn checkout -q "https://source.sakaiproject.org/svn/sakai/$S2_TAG" sakai
+    cd sakai/
+    REPO_REV=`svn info|grep Revision`
+    # SAK-17223 K2AuthenticationFilter
+    rm -rf login/
+    svn checkout -q https://source.sakaiproject.org/svn/login/branches/SAK-17223-2.7 login
+    # SAK-17222 K2UserDirectoryProvider
+    rm -rf providers
+    svn checkout -q https://source.sakaiproject.org/svn/providers/branches/SAK-17222-2.7 providers
+    # KERN-360 Servlet and TrustedLoginFilter RESTful services
+    cp -R $BUILD_DIR/sakai3/open-experiments/hybrid .
+    find hybrid -name pom.xml -exec perl -pwi -e 's/2\.8-SNAPSHOT/2\.7-SNAPSHOT/g' {} \;
+    perl -pwi -e 's/<\/modules>/<module>hybrid<\/module><\/modules>/gi' pom.xml
+    mvn -e clean install sakai:deploy -Dmaven.test.skip=true -Dmaven.tomcat.home=$BUILD_DIR/sakai2-demo
+    # configure sakai 2 instance
+    cd $BUILD_DIR
+    cp -f server.xml sakai2-demo/conf/server.xml 
+    echo "ui.service = trunk+SAK-17223+KERN-360 on HSQLDB" >> sakai2-demo/sakai/sakai.properties
+    echo "version.sakai = $REPO_REV" >> sakai2-demo/sakai/sakai.properties
+    echo "version.service = Built: $BUILD_DATE" >> sakai2-demo/sakai/sakai.properties
+    echo "serverName=nightly2.sakaiproject.org" >> sakai2-demo/sakai/sakai.properties
+    echo "webservices.allowlogin=true" >> sakai2-demo/sakai/sakai.properties
+    echo "webservice.portalsecret=nightly" >> sakai2-demo/sakai/sakai.properties
+    echo "samigo.answerUploadRepositoryPath= /tmp/sakai2-hybrid/" >> sakai2-demo/sakai/sakai.properties
+    # enable SAK-17223 K2AuthenticationFilter
+    echo "top.login=false" >> sakai2-demo/sakai/sakai.properties
+    echo "container.login=true" >> sakai2-demo/sakai/sakai.properties
+    echo "login.k2.authentication=true" >> sakai2-demo/sakai/sakai.properties
+    echo "login.k2.authentication.vaildateUrl=http://localhost:8008/var/cluster/user.cookie.json?c=" >> sakai2-demo/sakai/sakai.properties
+    # declare shared secret for trusted login from K2
+    echo "org.sakaiproject.util.TrustedLoginFilter.sharedSecret=e2KS54H35j6vS5Z38nK40" >> sakai2-demo/sakai/sakai.properties
+    date > $BUILD_DIR/sakai/.lastbuild
+fi
 
 # start sakai 2 tomcat
 echo "Starting sakai2 instance..."
-sakai2-demo/bin/startup.sh 
-
-# final cleanup
-cd $BUILD_DIR
-rm -rf ~/.m2/repository/org/sakaiproject
+cd $BUILD_DIR/sakai2-demo
+./bin/startup.sh
